add ForwardDiff norm of SVector workaround
